// libs/wpisa/feature-isa-journey/src/lib/core/models/personal-details-view-model.interface.ts
export interface PersonalDetailsViewModel {
  name: { label: string; value: string };
  dob: { label: string; value: string };
  profession: { label: string; value: string };
  nationality: { label: string; value: string };
  emailAddress: { label: string; value: string };
  phoneNumber: { label: string; value: string };
  homeAddress: { label: string; value: string };
}